[MOVE] 패키징 컨벤션 반영 (코리 반영)
