+++ app/src/main/java/org/sopt/dosopttemplate/presentation/SignUpActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.presentation
 
 import android.annotation.SuppressLint
 import android.content.Intent
@@ -7,13 +7,12 @@ import androidx.appcompat.app.AppCompatActivity
 
 import com.google.android.material.snackbar.Snackbar
 import android.widget.Toast
-import org.sopt.dosopttemplate.ServicePool.authService
+import org.sopt.dosopttemplate.module.ServicePool.authService
 //import com.codingmy.sopt_w1_hw1.databinding.ActivitySignupBinding
 import org.sopt.dosopttemplate.databinding.ActivitySignupBinding
-import org.sopt.dosopttemplate.retrofit2data.RequestLoginDto
-import org.sopt.dosopttemplate.retrofit2data.RequestSignUpDto
-import org.sopt.dosopttemplate.retrofit2data.ResponseLoginDto
-import org.sopt.dosopttemplate.retrofit2data.ResponseSignUpDto
+import org.sopt.dosopttemplate.presentation.login.LoginActivity
+import org.sopt.dosopttemplate.data.dto.follower.retrofit2data.RequestSignUpDto
+import org.sopt.dosopttemplate.data.dto.follower.retrofit2data.ResponseSignUpDto
 import retrofit2.Call
 import retrofit2.Response
 
@@ -37,12 +36,12 @@ class SignUpActivity : AppCompatActivity() {
                 signUp()
 
                 Toast.makeText(this, "회원가입 성공", Toast.LENGTH_SHORT).show()
-/*
-                //유저정보 -> 리스트로 구성
-                val userInfoList = UserInfoToListString()
+                /*
+                                //유저정보 -> 리스트로 구성
+                                val userInfoList = UserInfoToListString()
 
-                sendUserInfo(intent, userInfoList)
-*/
+                                sendUserInfo(intent, userInfoList)
+                */
                 //액티비티 이동
                 startActivity(intent)
 
@@ -60,38 +59,38 @@ class SignUpActivity : AppCompatActivity() {
     private fun checkCondition() =
         (binding.etSignupId.text.length < 11 && binding.etSignupId.text.length > 5) && (binding.etSignupPw.text.length < 13 && binding.etSignupPw.text.length > 7) && (binding.etSignupNickname.text.length > 0 && !binding.etSignupNickname.text.isBlank())
 
-/*
-    private fun sendUserInfo(
-        intent: Intent,
-        userInfoList: List<String>
-    ) {
-        //id pw 넘기기
-        intent.putStringArrayListExtra("userInfoList", ArrayList(userInfoList))
-        setResult(RESULT_OK, intent)
-    }
-*/
+    /*
+        private fun sendUserInfo(
+            intent: Intent,
+            userInfoList: List<String>
+        ) {
+            //id pw 넘기기
+            intent.putStringArrayListExtra("userInfoList", ArrayList(userInfoList))
+            setResult(RESULT_OK, intent)
+        }
+    */
 
-/*
-    private fun UserInfoToListString(): List<String> {
-        val userInfoList = listOf<String>(
-            binding.etSignupId.text.toString(),
-            binding.etSignupPw.text.toString(),
-            binding.etSignupNickname.text.toString(),
-            binding.etSignupMbti.text.toString()
-        )
-        return userInfoList
-    }
-*/
+    /*
+        private fun UserInfoToListString(): List<String> {
+            val userInfoList = listOf<String>(
+                binding.etSignupId.text.toString(),
+                binding.etSignupPw.text.toString(),
+                binding.etSignupNickname.text.toString(),
+                binding.etSignupMbti.text.toString()
+            )
+            return userInfoList
+        }
+    */
 
     private fun signUp() = with(binding) {
         val id = etSignupId.text.toString();
-        val pw =etSignupPw.text.toString()
-        val nickname=etSignupNickname.text.toString()
+        val pw = etSignupPw.text.toString()
+        val nickname = etSignupNickname.text.toString()
 
         btSignupButton.setOnClickListener {
             authService.signUp(RequestSignUpDto(id, pw, nickname))
                 .enqueue(object : retrofit2.Callback<ResponseSignUpDto> {
-                        override fun onResponse(
+                    override fun onResponse(
                         call: Call<ResponseSignUpDto>,
                         response: Response<ResponseSignUpDto>,
                     ) {
@@ -103,6 +102,7 @@ class SignUpActivity : AppCompatActivity() {
                             ).show()
                         }
                     }
+
                     override fun onFailure(call: Call<ResponseSignUpDto>, t: Throwable) {
                         Toast.makeText(this@SignUpActivity, "서버 에러 발생", Toast.LENGTH_SHORT).show()
                     }
